+++ kernel/stat.h
@@ -7,5 +7,5 @@ struct stat {
   uint ino;    // Inode number
   short type;  // Type of file
   short nlink; // Number of links to file
-  uint64 size; // Size of file in bytes
+  uint32 size; // Size of file in bytes
 };
